Math fix

+++ engine/include/genebits/engine/math/funcs_common.h
@@ -281,7 +281,7 @@ namespace rtmath
   ///
   inline long double Sqrt(long double x) noexcept
   {
-    return Sqrt(static_cast<float>(x));
+    return Sqrt(static_cast<double>(x));
   }
 
   ///
